Initial Commit of Fitting Stuff
Initial commit of LSQ fitting script and function. As it currently stands, it works but the fit is not ideal. This commit includes LSQfunc which calculates the LSQ value and then LSQModelFitting which is the script to do the actual fitting.

Things to improve: there is a rounding discrepancy in the x-axes and the final slopes do not currently get fit for. See comments in file for more details.