Make Red PM Read Optional in PiezoScan
Modified main PiezoScan script to include a flag that enables or disables the red power meter (pre-lens normalization) read. When the flag is false, the PM is not read and the mean&std values in the fits keywords are set to -9999. Also, in that case, the Normed cube is just divided by 1 (un-normed). Finally, a new keyword was added to state whether the flag was enabled or disabled.
If the flag is set to true, the code runs as it did before the mods (except the new keyword is still present).