Update analysis code to account for optional PM read
Updated the VFN_an_fitsNormed() function to account for the fact that the red PM value is now optional. This uses the pm flag keyword so that it is still backwards compatible. 

Also updated the header of VFN_An_getKwd() to reflect the fact that a blank array is returned when a non-existant keyword is requested. This behavior already existed but was not documented.