Modify norm in results printed by PiezoScan
Added an optinoal variable, pmNormReport, which defined what the assumed power is on the fiber tip for normalization in the printed values only. This does not affect the data cubes output by the code, it only affects the printed values and thus the values recorded in the DataNotes.txt files. This can be set to 1 to not normalize. The value is completely ignored when isPMNorm=true, ie. when the power meter is actually used.