Initial Commit of NullFinder Minimizer
This is the first, untested version of the auto null-finder script. It uses patternsearch to move the fiber to null point in a donut. It assumes the focus axis on the fiber has a Zaber and it does not move the vortex. It also does not save or output any data; it just returns the best null point and value.