Initial Commit of PiezoScanZabFoc
This is the same as the current PiezoScan code except it assumes the redPM zaber has been moved to the fiber focus axis s.t. the z-piezo is not used and the redPM is on a fixed stage. 
As such, this will do a regular 5D scan of the VFN setup (Fiber X, Fiber Y, Fiber Z, Vortex X, Vortex Y). This is tested and functional at this point. However, I haven't tried analyzing the output data yet. It should still be able to use the same analysis functions though.