Initial Commit of PI Stages Control Code
Code for controlling the PI Q-545 stages via an E-873 controller. This code was tested using the VFN_DMCharac_DMCompCode which was used to measure the step size of the stages with the Zygo. As such, that code serves as a good example of using this library. 
The BeginWithThis[...] is a modified version of the test script provided by PI.
The remaining functions are used for controlling the stages. NOTE: this does not follow the format used in the rest of the VFN code.