Rename setUp and cleanUp files in PIStage dir
Renamed the init and cleanUp files in the PI Stage control library to match the naming convention used by VFN and HCST. 

No other mods were made to the code in this commit so the code should still work as intended but I have not tested it with the rename. As such, only updates to code should be to rename the function calls in any scripts using this library.