KLS635 Laser Control Library
Add a new library for controlling the Thorlabs KLS635 library. This is based on the KPIC TLS.py python library. NOTE: the TLS library was originally written for a TLS laser, not the KLS. As such some functions behave a bit funny but the library works well enough for basic control from python and matlab. One known issue is that the library always things the KLS is disabled and it prints that the laser needs to be enabled before use. This is not actually the case; using enableOut() works just fine, the status bit seems to just be read incorrectly.